+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `@homebridge/hap-client` will be documented in this file. This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
+## BETA
+
+### Changed
+
+- updated dependencies
+
 ## v2.0.4 (2024-11-07)
 
 ### Changed