+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to `@homebridge/hap-client` will be documented in this file.
 ### Changed
 
 - updated dependencies
+- Added refresh of 'values' field when refreshServiceCharacteristics or getCharacteristic is executed
+- Added new setCharacteristic option, setCharacteristicByType which allows finding the characteristic by type.
 
 ## v2.0.4 (2024-11-07)