+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to `@homebridge/hap-client` will be documented in this file.
 - Added new setCharacteristic option, setCharacteristicByType which allows finding the characteristic by type.
 - Added new getResource request to retrieve snapshot images from camera's
 - Minor tweak to serviceName, and if the name is blank, use the name value from Accessory Information
+- Fixed issue of error handler triggering error when contacting a homebridge instance that is down
+- Added restart of monitor when client connections close
 
 ## v2.0.4 (2024-11-07)