+++ CHANGELOG.md
@@ -7,11 +7,11 @@ All notable changes to `@homebridge/hap-client` will be documented in this file.
 ### Changed
 
 - updated dependencies
-- Added refresh of 'values' field when refreshServiceCharacteristics or getCharacteristic is executed
-- Added new setCharacteristic option, setCharacteristicByType which allows finding the characteristic by type.
+- Added refresh of 'values' field when refreshServiceCharacteristics or getCharacteristic is used
+- Added new setCharacteristic option, setCharacteristicByType which finds characteristic by type for setting.
 - Added new getResource request to retrieve snapshot images from camera's
 - Minor tweak to serviceName, and if the name is blank, use the name value from Accessory Information
-- Fixed issue of error handler triggering error when contacting a homebridge instance that is down
+- Fixed issue of error handler triggering error when attempting to connect to a homebridge instance that is down
 - Added restart of monitor when client connections close
 
 ## v2.0.4 (2024-11-07)